fix(system): edit data type

+++ src/main/java/com/futurenbetter/saas/modules/order/entity/OrderItem.java
@@ -26,7 +26,7 @@ public class OrderItem {
     Long orderItemId;
 
     @Column(name = "unit_price")
-    Double unitPrice;
+    Long unitPrice;
 
     @Column(name = "order_item_status")
     @Enumerated(EnumType.STRING)
